Split functions into different files
